requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-Caching==2.0.2
Flask-Limiter==3.5.0
Flask-Migrate==4.0.4
Flask-Swagger-UI==4.11.1
celery==5.3.4
redis==5.0.1
pymongo==4.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
python-dateutil==2.8.2
email-validator==2.0.0
boto3==1.28.36
ebcli==4.0.18 